+++ buildSrc/src/main/kotlin/com/bugsnag/android/Versions.kt
@@ -9,16 +9,15 @@ object Versions {
     // Note minSdkVersion must be >=21 for 64 bit architectures
     val minSdkVersion = 14
     val compileSdkVersion = 30
-    val ndk = "16.1.4479499"
+    val ndk = "17.2.4988734"
     val java = JavaVersion.VERSION_1_7
     val kotlin = "1.3.72"
 
     // plugins
-    val androidGradlePlugin = "4.1.1"
-    val bintrayPlugin = "1.8.5"
-    val detektPlugin = "1.17.1"
-    val ktlintPlugin = "9.4.1"
-    val dokkaPlugin = "0.10.1"
+    val androidGradlePlugin = "7.0.2"
+    val detektPlugin = "1.18.1"
+    val ktlintPlugin = "10.1.0"
+    val dokkaPlugin = "1.5.0"
     val benchmarkPlugin = "1.0.0"
 
     // dependencies