+++ scripts/run-instrumentation-test.sh
@@ -39,7 +39,7 @@ fi
 echo "Android Tests [$(timestamp)]: Second app upload successful, url: $test_url"
 
 echo "Android Tests [$(timestamp)]: Starting test run"
-build_response=$(curl -X POST "https://api-cloud.browserstack.com/app-automate/espresso/build" -d \ "{\"devices\": $INSTRUMENTATION_DEVICES, \"app\": \"$app_url\", \"deviceLogs\" : true, \"testSuite\": \"$test_url\"}" -H "Content-Type: application/json" -u "$BROWSER_STACK_USERNAME:$BROWSER_STACK_ACCESS_KEY")
+build_response=$(curl -X POST "https://api-cloud.browserstack.com/app-automate/espresso/build" -d \ "{\"video\": false, \"singleRunnerInvocation\": \"$SINGLE_RUNNER\", \"devices\": $INSTRUMENTATION_DEVICES, \"app\": \"$app_url\", \"deviceLogs\" : true, \"testSuite\": \"$test_url\"}" -H "Content-Type: application/json" -u "$BROWSER_STACK_USERNAME:$BROWSER_STACK_ACCESS_KEY")
 
 build_id=$(echo "$build_response" | jq -r ".build_id")
 
@@ -58,7 +58,7 @@ status_response=$(curl -s -u "$BROWSER_STACK_USERNAME:$BROWSER_STACK_ACCESS_KEY"
 status=$(echo "$status_response" | jq -r ".status")
 
 WAIT_COUNT=0
-until [ "$status" == "\"done\"" ] || [ "$status" == "\"error\"" ] || [ "$status" == "\"failed\"" ] || [ $WAIT_COUNT -eq 100 ]; do
+until [ "$status" == "\"done\"" ] || [ "$status" == "\"error\"" ] || [ "$status" == "\"failed\"" ] || [ $WAIT_COUNT -eq 240 ]; do
     echo "Android Tests [$(timestamp)]: Current test status: $status, Time waited: $((WAIT_COUNT * 15))"
     ((WAIT_COUNT++))
     sleep 15